second commit

+++ app/products/models.py
@@ -1,7 +1,5 @@
-from sqlalchemy import Column, Integer, Date, String
-
+from sqlalchemy import Column, Integer, String, Numeric
 from app.database import Base
-
 
 class Products(Base):
     __tablename__ = 'products'
@@ -9,5 +7,5 @@ class Products(Base):
     id = Column(Integer, primary_key=True)
     title = Column(String)
     description = Column(String)
-    price = Column(Integer)
+    price = Column(Numeric(10, 2))
     amount = Column(Integer)
