+++ app/products/schemas.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from pydantic import BaseModel
 
 
@@ -5,7 +6,7 @@ class SProduct(BaseModel):
     id: int
     title: str
     description: str
-    price: int
+    price: Decimal
     amount: int
 
     class Config:
